move is_active from prover_client to prover_node
